refactor: use stdbool, size_t, clear some warnings

+++ examples/array.c
@@ -4,8 +4,6 @@
  */
 
 #include <stdio.h>
-#include <stdint.h>
-#include <inttypes.h>
 #include <string.h>
 #include <errno.h>
 #include <stdbool.h>
@@ -28,26 +26,26 @@ static char *names[] =
     NULL,
 };
 
-static uintmax_t
-mustache_read_file(mustache_api_t *api, void *data, char *buf, uintmax_t sz)
+static size_t
+mustache_read_file(mustache_api_t *api, void *data, char *buf, size_t sz)
 {
     /* Do whatever needs to be done to get sz number of bytes into buf.
      */
     return fread(buf, sizeof(char), sz, template);
 }
 
-static uintmax_t
+static size_t
 mustache_write_stdout(mustache_api_t *api, void *data,
-                      char const *buf, uintmax_t sz)
+                      char const *buf, size_t sz)
 {
     /* Do whatever it takes to write sz number of bytes from buf to somewhere.
      */
     return fwrite(buf, sizeof(char), sz, stdout);
 }
 
-static uintmax_t
+static size_t
 mustache_write_null(mustache_api_t *api, void *data,
-                    char const *buf, uintmax_t sz)
+                    char const *buf, size_t sz)
 {
     /* Fake success
      */
@@ -56,12 +54,12 @@ mustache_write_null(mustache_api_t *api, void *data,
 
 static void
 mustache_error(mustache_api_t *api, void *data,
-               uintmax_t line, char const *error)
+               size_t line, char const *error)
 {
-    fprintf(stderr, "error in template: %" PRIu64  ": %s\n", line, error);
+    fprintf(stderr, "error in template: %zu: %s\n", line, error);
 }
 
-static uintmax_t
+static size_t
 mustache_sectget(mustache_api_t *api, void *data, mustache_token_section_t *s)
 {
     if (strcmp(s->name, "members") == 0) {
@@ -83,7 +81,7 @@ mustache_sectget(mustache_api_t *api, void *data, mustache_token_section_t *s)
     return false;
 }
 
-static uintmax_t
+static size_t
 mustache_varget(mustache_api_t *api, void *data, mustache_token_variable_t *t)
 {
     if (strcmp(t->text, "name") == 0) {
